refactor: Channel, DM, Sidebar, Header, Chatroom Container 삭제
- react-redux의 useSelector와 useDispatch를 사용하면서 해당 Container 삭제
