+++ client/src/components/molecules/index.ts
@@ -1,6 +1,6 @@
 import { ActiveProfileImg } from './ActiveProfileImg/ActiveProfileImg';
-import Channel from './Channel/ChannelContainer';
-import DM from './DM/DMContainer';
+import { Channel } from './Channel/Channel';
+import { DM } from './DM/DM';
 import { Section } from './Section/Section';
 import { InputMessage } from './InputMessage/InputMessage';
 import { SendMessageButton } from './SendMessageButton/SendMessageButton';
