--- client/src/components/organisms/Header/HeaderContainer.tsx
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux';
-import { Header } from './Header';
-
-function mapReduxStateToReactProps(state: any) {
-  return {
-    profileUri: state.userData.profileUri
-  };
-}
-
-export default connect(mapReduxStateToReactProps)(Header);
