+++ client/src/components/organisms/index.ts
@@ -1,5 +1,5 @@
-import Header from './Header/HeaderContainer';
-import Sidebar from './Sidebar/SidebarContainer';
+import { Header } from './Header/Header';
+import { Sidebar } from './Sidebar/Sidebar';
 import { ChatroomHeader } from './ChatroomHeader/ChatroomHeader';
 import { ChatroomBody } from './ChatroomBody/ChatroomBody';
 import { LoginForm } from './LoginForm/LoginForm';
